Converting ServiceManager to interface
