+++ android-agent/src/main/java/io/opentelemetry/android/internal/services/ServiceManagerImpl.kt
@@ -1,0 +1,49 @@
+/*
+ * Copyright The OpenTelemetry Authors
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+package io.opentelemetry.android.internal.services
+
+import io.opentelemetry.android.internal.services.periodicwork.PeriodicWorkService
+import java.util.Collections
+
+internal class ServiceManagerImpl(services: List<Any>) : ServiceManager {
+    private val services: Map<Class<out Any>, Any>
+
+    init {
+        val map: MutableMap<Class<out Any>, Any> = HashMap()
+        for (service in services) {
+            map[service.javaClass] = service
+        }
+        this.services = Collections.unmodifiableMap(map)
+    }
+
+    override fun getPreferencesService(): PreferencesService {
+        return getService(PreferencesService::class.java)
+    }
+
+    override fun getCacheStorageService(): CacheStorageService {
+        return getService(CacheStorageService::class.java)
+    }
+
+    override fun getPeriodicWorkService(): PeriodicWorkService {
+        return getService(PeriodicWorkService::class.java)
+    }
+
+    override fun start() {
+        for (service in services.values) {
+            if (service is Startable) {
+                service.start()
+            }
+        }
+    }
+
+    @Suppress("UNCHECKED_CAST")
+    private fun <T : Any> getService(type: Class<T>): T {
+        val service =
+            services[type]
+                ?: throw IllegalArgumentException("Service not found: $type")
+        return service as T
+    }
+}
